+++ client/components/tasks/TaskProperties.vue
@@ -6,7 +6,7 @@
   <select-date @select="onSelectDueDate" :active.sync="showSelectDueDate"></select-date>
   <select-date @select="onSelectStartDate" :active.sync="showSelectStartDate"></select-date>
   
-  <v-subheader>Responsabilités</v-subheader>
+  <v-subheader>{{ $t('Duties') }}</v-subheader>
   <v-list class="elevation-1">
     <v-list-tile @click="showChooseAssignedToDialog = true">
       <v-list-tile-avatar :color="isOnline(task.assignedTo)">
@@ -16,7 +16,7 @@
       <v-list-tile-title>
         <span v-show="task.assignedTo">Assignée à </span>
         <span>{{ formatUser(task.assignedTo) }}</span>
-        <span v-show="!task.assignedTo">Non assignée </span>
+        <span v-show="!task.assignedTo">{{ $t('Unassigned')}}</span>
       </v-list-tile-title>
       </v-list-tile-content>
       <v-list-tile-action>
@@ -34,10 +34,10 @@
           <v-icon>calendar_today</v-icon>
         </v-list-tile-avatar>
         <v-list-tile-content>
-          <v-list-tile-title>Date de début</v-list-tile-title>
+          <v-list-tile-title>{{ $t('Start date') }}</v-list-tile-title>
           <v-list-tile-sub-title>
             <span v-show="task.startDate">{{ formatDate(task.startDate) }}</span>
-            <span v-show="!task.startDate">Aucune</span>
+            <span v-show="!task.startDate">{{ $t('None') }}</span>
           </v-list-tile-sub-title>
         </v-list-tile-content>
         <v-list-tile-action>
@@ -54,10 +54,10 @@
           <v-icon>alarm_on</v-icon>
         </v-list-tile-avatar>
         <v-list-tile-content>
-          <v-list-tile-title>Date de fin</v-list-tile-title>
+          <v-list-tile-title>{{ $t('End date') }}</v-list-tile-title>
           <v-list-tile-sub-title>
             <span v-show="task.dueDate">{{ formatDate(task.dueDate) }}</span>
-            <span v-show="!task.dueDate">Aucune</span>
+            <span v-show="!task.dueDate">{{ $t('None') }}</span>
           </v-list-tile-sub-title>
         </v-list-tile-content>
         <v-list-tile-action>