+++ imports/api/projects/server/publications.js
@@ -170,9 +170,10 @@ publishComposite("project", function(projectId) {
           const members = Array.from(project.members) || [];
           const tasks = Tasks.find({ projectId: project._id, deleted: {$ne: true} }, {fields: { createdBy: 1, updatedBy: 1}})
           tasks.map(task => {
-            if (task.createdBy) members.push(task.createdBy)
-            if (task.updatedBy) members.push(task.updatedBy)
-          })
+            if (task.createdBy) members.push(task.createdBy);
+            if (task.updatedBy) members.push(task.updatedBy);
+            if (task.watchers) members.push(task.watchers.map(wacher => { return watcher; }));
+          });
 
           if (project.organizationId) {
             const organization = Organizations.findOne({ _id: project.organizationId }, { fields: {members: 1}});