+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+
+### Changed
+
+### Removed
+
+## [1.8.0] - 2019-08-03
+### Added
 - More i18n translations
 - Edit list name right after creation
 - Task reminder